Factor common CLI task code to Gemstash::CLI::Base
