+++ lib/gemstash/cli/start.rb
@@ -5,17 +5,9 @@ module Gemstash
   class CLI
     # This implements the command line start task to start the Gemstash server:
     #  $ gemstash start
-    class Start
-      include Gemstash::Env::Helper
-
-      def initialize(cli)
-        Gemstash::Env.current = Gemstash::Env.new
-        @cli = cli
-      end
-
+    class Start < Gemstash::CLI::Base
       def run
-        check_rubygems_version
-        store_config
+        prepare
         setup_logging
         store_daemonized
         Puma::CLI.new(args, Gemstash::Logging::StreamLogger.puma_events).run
@@ -28,20 +20,8 @@ module Gemstash
         Gemstash::Logging.setup_logger(gemstash_env.base_file("server.log"))
       end
 
-      def store_config
-        config = Gemstash::Configuration.new(file: @cli.options[:config_file])
-        gemstash_env.config = config
-      end
-
       def store_daemonized
         Gemstash::Env.daemonized = daemonize?
-      end
-
-      def check_rubygems_version
-        @cli.say(@cli.set_color("Rubygems version is too old, " \
-                                 "please update rubygems by running: " \
-                                 "gem update --system", :red)) unless
-        Gem::Requirement.new(">= 2.4").satisfied_by?(Gem::Version.new(Gem::VERSION))
       end
 
       def daemonize?
@@ -58,10 +38,6 @@ module Gemstash
 
       def config_args
         ["--config", puma_config]
-      end
-
-      def pidfile_args
-        ["--pidfile", gemstash_env.base_file("puma.pid")]
       end
 
       def daemonize_args
