+++ lib/gemstash/cli/status.rb
@@ -5,28 +5,16 @@ module Gemstash
   class CLI
     # This implements the command line status task to check the server status:
     #  $ gemstash status
-    class Status
-      include Gemstash::Env::Helper
-
-      def initialize(cli)
-        Gemstash::Env.current = Gemstash::Env.new
-        @cli = cli
-      end
-
+    class Status < Gemstash::CLI::Base
       def run
-        store_config
+        prepare
         Puma::ControlCLI.new(args).run
       end
 
     private
 
-      def store_config
-        config = Gemstash::Configuration.new(file: @cli.options[:config_file])
-        gemstash_env.config = config
-      end
-
       def args
-        ["--pidfile", gemstash_env.base_file("puma.pid"), "status"]
+        pidfile_args + %w(status)
       end
     end
   end
